refactor: setInvoiceNumber
