+++ src/Service/OrderManager.php
@@ -7,6 +7,7 @@ namespace Linio\SellerCenter\Service;
 use Linio\SellerCenter\Factory\Xml\Order\OrderItemsFactory;
 use Linio\SellerCenter\Model\Order\OrderItem;
 use Linio\SellerCenter\Transformer\Order\OrderItemsTransformer;
+use Linio\SellerCenter\Response\SuccessResponse;
 
 class OrderManager extends BaseOrderManager
 {
@@ -41,5 +42,36 @@ class OrderManager extends BaseOrderManager
         );
 
         return $orderItems;
+    }
+
+    public function setInvoiceNumber(
+        int $orderItemId,
+        string $invoiceNumber
+    ): SuccessResponse {
+        $action = 'SetInvoiceNumber';
+        $parameters = $this->makeParametersForAction($action);
+
+        $parameters->set([
+            'OrderItemId' => $orderItemId,
+            'InvoiceNumber' => $invoiceNumber,
+        ]);
+
+        $requestId = $this->generateRequestId();
+        $response = $this->executeAction(
+            $action,
+            $parameters,
+            $requestId,
+            'POST'
+        );
+
+        $this->logger->info(
+            sprintf(
+                '%d::%s::APIResponse::SellerCenterSdk: Invoice Number Set',
+                $requestId,
+                $action
+            )
+        );
+
+        return $response;
     }
 }
